Moves enumerations base to separate file
